# qa/openapi-launch-contract.yaml
openapi: 3.1.0
info:
  title: AI Job Application Copilot Launch Contract
  version: 0.1.0
servers:
  - url: http://localhost:3000
security:
  - WebSession: []
components:
  securitySchemes:
    WebSession:
      type: apiKey
      in: cookie
      name: session
    ExtensionBearer:
      type: http
      scheme: bearer
...
paths:
...
  /api/resume/upload:
    post:
      summary: Upload a resume with strict type and size validation
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200": { description: Resume parsed }
        "400": { description: Invalid file }
        "413": { description: File too large }
        "401": { description: Missing web session }
...